ci: Jenkinsfile: Set TMPDIR for the Docker agent
The Docker container for foreign architectures is ran with qemu-user
that needs a writable TMPDIR for syscall emulation. Otherwise syscalls
may fail.

The Jenkins at emlix sets a global TMPDIR=/data/tmp, but this directory
does not exist in the container. Set a well-known path that is writable
in the container.